Created admin.dashboard.loan_management.open_loans.ItemTemplate3; Edited admin.dashboard.loan_management.open_loans; Deleted admin.dashboard.loan_management.open_loans.RowTemplate13 and admin.dashboard.loan_management.open_loans.ItemTemplate29
