--- client_code/admin/dashboard/loan_management/open_loans/ItemTemplate29/form_template.yaml
@@ -1,3 +1,0 @@
-is_package: true
-container: {type: ColumnPanel}
-components: []
