+++ client_code/admin/dashboard/loan_management/open_loans/ItemTemplate3/form_template.yaml
@@ -1,0 +1,136 @@
+is_package: true
+container: {type: ColumnPanel}
+components:
+- name: xy_panel_1
+  properties: {height: 441, role: outlined-card}
+  type: XYPanel
+  layout_properties: {grid_position: 'IMFITS,VYSBGC'}
+  components:
+  - name: label_3
+    properties: {text: '---------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------',
+      foreground: grey, align: center}
+    type: Label
+    layout_properties: {x: 0.147705078125, y: 108.6477323025465, width: 938.4028450195312}
+  - name: image_1
+    properties: {height: 63.399932861328125, source: 'https://freesvg.org/img/abstract-user-flat-4.png'}
+    type: Image
+    layout_properties: {x: 1.6363664120435715, y: 60.409094378352165, width: 66}
+  - name: borrower_mobile
+    properties: {}
+    type: Label
+    layout_properties: {x: 77.17046980559826, y: 85.48864416778088, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''borrower_mobile'']', writeback: false}
+  - name: borrower_name
+    properties: {bold: true}
+    type: Label
+    layout_properties: {x: 77.06814436614513, y: 59.68181852996349, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''borrower_full_name'']', writeback: false}
+  - name: image_1_copy
+    properties: {height: 63.399932861328125, source: 'https://freesvg.org/img/abstract-user-flat-4.png'}
+    type: Image
+    layout_properties: {x: 853.3750152587891, y: 53.13636641204357, width: 85.17045593261719}
+  - name: product_details
+    properties: {align: center}
+    type: Label
+    layout_properties: {x: 350.1818528622389, y: 21.704544588923454, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''product_description'']', writeback: false}
+  - name: Product
+    properties: {bold: true, align: center}
+    type: Label
+    layout_properties: {x: 343.5568528622389, y: 0.04545454680919647, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''product_name'']', writeback: false}
+  - name: lendor_name
+    properties: {bold: true, align: right}
+    type: Label
+    layout_properties: {x: 669.7625396735966, y: 51.38750457763672, width: 178.9869755859375}
+    data_bindings:
+    - {property: text, code: 'self.item[''lender_full_name'']', writeback: false}
+  - name: lendor_mobile
+    properties: {align: right}
+    type: Label
+    layout_properties: {x: 655.5124176032841, y: 75.1875, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''lender_mobile'']', writeback: false}
+  - name: outlined_button_1
+    properties: {text: View Loan Details, role: outlined-button, align: center, font_size: null,
+      background: '#05264d', foreground: white, bold: true}
+    type: Button
+    layout_properties: {x: -3.1072486572265916, y: 378.856721494751, width: 938.393}
+    event_bindings: {click: outlined_button_1_click}
+  - name: return
+    properties: {text: '     ------Not Invested Yet-----', italic: true, bold: false,
+      align: center}
+    type: Label
+    layout_properties: {x: 3.4642791748046875, y: 316.1964359283447, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''loan_disbursed_timestamp'']', writeback: false}
+  - name: label_9
+    properties: {text: 'Invested on : ', bold: true}
+    type: Label
+    layout_properties: {x: 20.803573608398438, y: 270.1964359283447, width: 103.589}
+  - name: total
+    properties: {align: right}
+    type: Label
+    layout_properties: {x: 724.8928680419922, y: 300.1785831451416, width: 94.5536}
+  - name: label_4
+    properties: {text: Rs, bold: true}
+    type: Label
+    layout_properties: {x: 842.4643402099609, y: 300.1785831451416, width: 22.3929}
+  - name: label_2
+    properties: {text: 'Overall Return :', bold: true}
+    type: Label
+    layout_properties: {x: 739.8928680419922, y: 270.1964359283447, width: 200}
+  - name: label_8
+    properties: {text: 'Principal Amount :'}
+    type: Label
+    layout_properties: {x: 23.089279174804688, y: 232.5535831451416, width: 200}
+  - name: label_7
+    properties: {text: 'Term :'}
+    type: Label
+    layout_properties: {x: 23.089279174804688, y: 197.1785831451416, width: 200}
+  - name: term
+    properties: {align: left}
+    type: Label
+    layout_properties: {x: 739.8928680419922, y: 197.1785831451416, width: 94.5}
+    data_bindings:
+    - {property: text, code: 'self.item[''tenure'']', writeback: false}
+  - name: principal_amount
+    properties: {}
+    type: Label
+    layout_properties: {x: 739.8928680419922, y: 232.5535831451416, width: 200}
+    data_bindings:
+    - {property: text, code: 'self.item[''loan_amount'']', writeback: false}
+  - name: interest_rate
+    properties: {align: left, text: ''}
+    type: Label
+    layout_properties: {x: 736.1607624755859, y: 165.1785831451416, width: 30.5714}
+    data_bindings:
+    - {property: text, code: 'self.item[''interest_rate'']', writeback: false}
+  - name: label_1
+    properties: {text: '%', align: left, bold: false}
+    type: Label
+    layout_properties: {x: 776.7321624755859, y: 165.1785831451416, width: 30.3929}
+  - name: label_6
+    properties: {text: 'Interest Rate :'}
+    type: Label
+    layout_properties: {x: 23.089279174804688, y: 165.1785831451416, width: 200}
+  - name: label_5
+    properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
+      visible: true, text: Loan id, font_size: null, font: '', spacing_above: small,
+      icon_align: left, spacing_below: small, italic: false, background: '', bold: false,
+      underline: false, icon: ''}
+    type: Label
+    layout_properties: {x: 23.053573608398438, y: 131.48214149475098, width: 160}
+  - name: loan_id
+    properties: {role: null, align: left, tooltip: '', border: '', foreground: '',
+      visible: true, text: '', font_size: null, font: '', spacing_above: small, icon_align: left,
+      spacing_below: small, italic: false, background: '', bold: false, underline: false,
+      icon: ''}
+    type: Label
+    layout_properties: {x: 736.1607513427734, y: 135.1785831451416, width: 160}
+    data_bindings:
+    - {property: text, code: 'self.item[''loan_id'']'}
