--- client_code/admin/dashboard/loan_management/open_loans/RowTemplate13/__init__.py
@@ -1,22 +1,0 @@
-from ._anvil_designer import RowTemplate13Template
-from anvil import *
-import anvil.server
-import anvil.google.auth, anvil.google.drive
-from anvil.google.drive import app_files
-import anvil.users
-import anvil.tables as tables
-import anvil.tables.query as q
-from anvil.tables import app_tables
-
-class RowTemplate13(RowTemplate13Template):
-  def __init__(self, **properties):
-    # Set Form properties and Data Bindings.
-    self.init_components(**properties)
-
-    # Any code you write here will run before the form opens.
-
-  def link_1_click(self, **event_args):
-    """This method is called when the link is clicked"""
-    value_to_pass = self.label_1.text
-    open_form('admin.dashboard.loan_management.open_loans.view_profile_4', value_to_pass)
-
