--- client_code/admin/dashboard/loan_management/open_loans/RowTemplate13/form_template.yaml
@@ -1,38 +1,0 @@
-container: {type: DataRowPanel}
-components:
-- type: Label
-  properties: {}
-  name: label_2
-  layout_properties: {column: WMIWXU}
-  data_bindings:
-  - {property: text, code: 'self.item[''coustmer_id'']'}
-- type: Label
-  properties: {}
-  name: label_3
-  layout_properties: {column: PVBYZX}
-  data_bindings:
-  - {property: text, code: 'self.item[''full_name'']'}
-- type: Label
-  properties: {role: null, align: left, tooltip: '', border: '', foreground: green,
-    visible: true, text: '', font_size: null, font: '', spacing_above: small, icon_align: left,
-    spacing_below: small, italic: false, background: '', bold: false, underline: false,
-    icon: ''}
-  name: label_4
-  layout_properties: {column: HCAQYD}
-  data_bindings:
-  - {property: text, code: 'self.item[''loan_status'']'}
-- type: Link
-  properties: {role: null, url: '', align: left, tooltip: '', border: '', foreground: '',
-    visible: true, text: view profile, font_size: null, wrap_on: mobile, font: '',
-    col_spacing: medium, spacing_above: small, icon_align: left, col_widths: '', spacing_below: small,
-    italic: false, background: '', bold: false, underline: false, icon: ''}
-  name: link_1
-  layout_properties: {column: DBUUUO}
-  event_bindings: {click: link_1_click}
-- type: Label
-  properties: {}
-  name: label_1
-  layout_properties: {column: EPCWQG}
-  data_bindings:
-  - {property: text, code: 'self.item[''loan_id'']'}
-is_package: true
