+++ client_code/admin/dashboard/loan_management/open_loans/form_template.yaml
@@ -12,47 +12,10 @@ components:
     properties: {}
     name: home_button_admin_1
     layout_properties: {grid_position: 'ADUYHR,SVOLIL', full_width_row: true}
-  - type: DataGrid
-    properties:
-      role: null
-      columns:
-      - {id: EPCWQG, title: Loan ID, data_key: column_1, $$hashKey: 'object:53723'}
-      - {id: WMIWXU, title: Customer ID, data_key: column_2, $$hashKey: 'object:53724'}
-      - {id: PVBYZX, title: Full Name, data_key: column_3, $$hashKey: 'object:53725'}
-      - {id: HCAQYD, title: Loan Status, data_key: column_4, $$hashKey: 'object:53773'}
-      - {id: DBUUUO, title: '', data_key: column_5, $$hashKey: 'object:53775'}
-      auto_header: true
-      tooltip: ''
-      border: ''
-      foreground: ''
-      rows_per_page: 20
-      visible: false
-      wrap_on: never
-      show_page_controls: true
-      spacing_above: large
-      spacing_below: small
-      background: ''
-    name: data_grid_1
-    layout_properties:
-      col_widths: {}
-      grid_position: HOLCPI,UNGSMN
-    components:
-    - type: RepeatingPanel
-      properties: {spacing_above: none, spacing_below: none, item_template: admin.dashboard.loan_management.open_loans.RowTemplate13}
-      name: repeating_panel_1
-      layout_properties: {}
-    - type: Link
-      properties: {role: null, url: '', align: right, tooltip: '', border: '', foreground: '',
-        visible: true, text: '', font_size: null, wrap_on: mobile, font: '', col_spacing: medium,
-        spacing_above: small, icon_align: left, col_widths: '', spacing_below: small,
-        italic: false, background: '', bold: false, underline: false, icon: 'fa:angle-double-left'}
-      name: link_2
-      layout_properties: {}
-      event_bindings: {click: link_2_click}
-  - name: repeating_panel_2
-    properties: {item_template: bank_users.repeating_panels.admin}
+  - name: repeating_panel_1
+    properties: {item_template: admin.dashboard.loan_management.open_loans.ItemTemplate3}
     type: RepeatingPanel
-    layout_properties: {grid_position: 'PFFQKX,WVVVCF'}
+    layout_properties: {grid_position: 'INEUUA,WYYJYD'}
 - type: FlowPanel
   properties: {}
   name: navbar_links
